E-20: extract FunctionEmitter from ReflectionMetadataEmitter (#1361)
Extracts the function-emission band (~1,540 lines) of
ReflectionMetadataEmitter into `internal sealed class FunctionEmitter`.
This is the last Wave-1 emitter collaborator extraction before the E-21
forwarder cleanup.

Band moved (verbatim bodies, mechanical rewrites only):
- EmitFunction (the ordinary managed-body MethodDef emitter) + its
  method-level NullableContextAttribute compaction ChooseMethodNullableContext
  (issue #834)
- the two interop shapes: EmitPInvokeFunction (@DllImport PinvokeImpl,
  ADR-0086 / issue #727) and EmitLibraryImportFunction (@LibraryImport
  managed-stub-plus-blittable-inner pair, ADR-0092 / issue #758) with
  their EmitLibraryImportOuterBody / EmitFieldMarshalRow /
  MapPInvokeImportAttributes helpers
- EmitGlobalFieldDefs (top-level global FieldDef emitter, issue #191)
- BuildMoveNextBodyBytes (async state-machine MoveNext body-bytes callback
  StateMachineEmitter drives via an injected Func)

Boundary helpers that stay on the root (shared with the EmitCore driver,
reached from FunctionEmitter through `outer`):
- EncodeAsyncReturnType — already internal (E-19); part of the async
  return-type band, also consumed by UserTokenResolver
- EmitExtensionAttribute — widened private -> internal; also stamps the
  <Program> host TypeDef via EmitCore's EmitProgramExtensionMarker

Wiring:
- ctor (ReflectionMetadataEmitter outer) back-ref; convenience fields read
  emitCtx / cache / remaps / signatures / memberRefs off outer at
  construction time. The later-built wellKnown / closures / stateMachines /
  customAttrEncoder peers are reached through outer (they are constructed in
  EmitCore, after this collaborator).
- constructed `internal readonly FunctionEmitter functions;` in the private
  ctor after userTokens (it only reads the earlier-built context objects).
- customAttrEncoder widened private -> internal for the outer back-reference.
- MethodBodyEmitSession sessions built as
  `new ReflectionMetadataEmitter.MethodBodyEmitSession(this.outer, il)`.

Forwarders kept for E-21 (function-emission band still called from EmitCore
and StateMachineEmitter): EmitFunction (EmitCore's function loop + a
ClosureEmitter method-group binding), EmitGlobalFieldDefs (one EmitCore call),
BuildMoveNextBodyBytes (StateMachineEmitter Func callback binding).

Verification: dotnet build GSharp.sln -c Release -graph → 0 warnings,
0 errors. PE byte-diff: 119/119 samples byte-for-byte identical vs the
pre-E-20 baseline (3 GsharpExtensions* skipped), including the PInvoke /
LibraryImport and async/generic samples that exercise the moved bodies.
